+++ src/views/home/footer/index.ts
@@ -3,7 +3,6 @@ export type { FooterProps } from "./footer";
 export type {
   FooterColumn,
   FooterContent,
-  FooterImage,
   FooterLink,
   FooterNewsletter,
 } from "./footer.types";